refactor: 💡 place each component in individual folders
