+++ libs/react/src/lib/alert/alert.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Canvas } from '@storybook/addon-docs'
 import Alert from './alert'
-import { Button } from '../form'
+import { Button } from '../form/button/button'
 
 export const Template = ({ children, ...props }) => (
   <Alert {...props}>{children}</Alert>
