+++ libs/react/src/lib/card/card.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Canvas } from '@storybook/addon-docs'
 import Card from './card'
-import Button from '../form/button'
+import Button from '../form/button/button'
 
 export const Template = ({ children, ...props }) => (
   <Card {...props}>{children}</Card>
