+++ libs/react/src/lib/form/buttonGroup/buttonGroup.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Canvas } from '@storybook/addon-docs'
 import { ButtonGroup } from './buttonGroup'
-import { Button } from './button'
+import { Button } from '../button/button'
 
 export const Template = ({ children, ...props }) => (
   <ButtonGroup {...props}>
