+++ libs/react/src/lib/form/buttonGroup/buttonGroup.test.tsx
@@ -1,10 +1,12 @@
 import { render, screen } from '@testing-library/react'
-import { Button, ButtonGroup } from './'
+import { Button, ButtonGroup } from '..'
 
 describe('ButtonGroup', () => {
   it('renders single button', () => {
     render(
-      <ButtonGroup><Button>Hello</Button></ButtonGroup>
+      <ButtonGroup>
+        <Button>Hello</Button>
+      </ButtonGroup>
     )
 
     expect(screen.getAllByRole('button')).toHaveLength(1)
