+++ libs/react/src/lib/form/buttonGroup/buttonGroup.tsx
@@ -1,7 +1,7 @@
 import { ReactElement, useEffect, useState } from 'react'
 import { ButtonVariant } from '@sebgroup/extract'
-import { Button, ButtonProps } from './button'
-import { Group } from '../layout'
+import { Button, ButtonProps } from '../button/button'
+import { Group } from '../../layout/group/group'
 
 interface ButtonGroupProps {
   children: ReactElement<ButtonProps> | ReactElement<ButtonProps>[]
@@ -9,12 +9,18 @@ interface ButtonGroupProps {
   variant?: ButtonVariant
 }
 type Props = ButtonProps & JSX.IntrinsicAttributes
-export const ButtonGroup = ({ children, selectedIndex, variant }: ButtonGroupProps) => {
+export const ButtonGroup = ({
+  children,
+  selectedIndex,
+  variant,
+}: ButtonGroupProps) => {
   const [selected, setSelected] = useState(selectedIndex)
   const [buttons, setButtons] = useState<Props[]>([])
 
   useEffect(() => {
-    const buttonProps: Props[] = ((children instanceof Array) ? children : [children])
+    const buttonProps: Props[] = (
+      children instanceof Array ? children : [children]
+    )
       .filter((b) => b && (b as ReactElement).props)
       .map((b, ix) => {
         const bp = (b as ReactElement).props
@@ -22,17 +28,16 @@ export const ButtonGroup = ({ children, selectedIndex, variant }: ButtonGroupPro
           ...bp,
           variant,
           key: bp.key || `btn_${ix}`,
-          active: (ix === selected),
+          active: ix === selected,
           onClick: (e) => {
             setSelected(ix)
             if (bp.onClick) bp.onClick(e)
-          }
+          },
         }
         return props
       })
     setButtons(buttonProps)
   }, [children, selected, variant])
-  
 
   return (
     <Group>
