+++ libs/react/src/lib/form/index.ts
@@ -1,5 +1,5 @@
-export * from './button'
-export * from './buttonGroup'
+export * from './button/button'
+export * from './buttonGroup/buttonGroup'
 export * from './form'
-export * from './input'
-export * from './text'
+export * from './input/input'
+export * from './text/text'
