+++ libs/react/src/lib/form/input/input.stories.mdx
@@ -1,6 +1,5 @@
 import { Meta, Story, Canvas } from '@storybook/addon-docs'
 import { TextInput, EmailInput, NumberInput } from './input'
-import { Form } from './form'
 
 export const TextInputTemplate = ({ ...props }) => <TextInput {...props} />
 export const EmailInputTemplate = ({ ...props }) => <EmailInput {...props} />
