+++ libs/react/src/lib/layout/flexbox/flexbox.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Canvas } from '@storybook/addon-docs'
-import { Flexbox } from './'
-import { Button } from '../form'
+import { Flexbox } from '../'
+import { Button } from '../../form'
 
 export const Template = (props) => (
   <Flexbox {...props}>
@@ -10,7 +10,7 @@ export const Template = (props) => (
   </Flexbox>
 )
 
-<Meta title="Components/Flexbox" component={Flexbox} />
+<Meta title="Components/Layout/Flexbox" component={Flexbox} />
 
 # Flexbox
