+++ libs/react/src/lib/layout/group/group.stories.mdx
@@ -1,12 +1,12 @@
 import { Meta, Story, Canvas } from '@storybook/addon-docs'
-import { Group } from './'
-import { TextInput, Button } from '../form'
+import { Group } from '../'
+import { TextInput, Button } from '../../form'
 
 export const Template = ({ children, ...props }) => (
   <Layout {...props}>{children}</Layout>
 )
 
-<Meta title="Components/Layout" component={Layout} />
+<Meta title="Components/Layout/Group" component={Layout} />
 
 # Layouts
