+++ libs/react/src/lib/modal/modal.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story, Canvas } from '@storybook/addon-docs'
 import Modal from './modal'
 import ModalSample from './modalSample'
-import Button from '../form/button'
+import Button from '../form/button/button'
 
 export const Template = ({ children, ...props }) => (
   <Modal {...props}>{children}</Modal>
